requirements.txt
customtkinter==5.2.1
imagesize==1.4.1
matplotlib==3.7.2
numpy==2.2.6
opencv_python==4.9.0.80
pandas==2.2.3
Pillow==10.0.1
pycocotools==2.0.7
pyduinocli==0.35.0
pygame==2.5.2
pyserial==3.5
segmentation_models_pytorch==0.1.3
setuptools==68.0.0
simple_pid==2.0.0
torch==2.2.1+cu118
tqdm==4.66.2
typing_extensions==4.13.2
